drf-requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==4.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coverage==6.4.1
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-dotenv==1.4.2
django-etc==1.3.1
django-mptt==0.13.4
django-rest-knox==4.2.0
django-social-share==2.2.1
django-summernote==0.8.20.0
django-taggit==3.0.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.13.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
http-ece==1.1.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
iniconfig==1.1.1
lxml==4.9.0
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==0.4.3
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==21.3
path==16.4.0
pathlib==1.0.1
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pyflakes==2.4.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pymemcache==3.5.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.0.0
PyPDF3==1.0.6
pytest==7.1.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-html==3.1.1
pytest-metadata==2.0.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.1.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
